Refactor responsiveness to include a new option
